+++ src/Territory.h
@@ -24,65 +24,65 @@
 class Territory {
  public:
   int getId() { return id; }
-  float getArea() { return area; }
+  double getArea() { return area; }
   int getWidth() { return width; }
   int getLength() { return length; }
-  float getDensity() { return density; }
+  double getDensity() { return density; }
   int getPopulation() { return population; }
-  float getGas() { return gas; }
-  float getSewage() { return sewage; }
-  std::vector<float> getAgeProp() { return ageProp; }
-  float getMaleProp() { return maleProp; }
-  float getHighEdProp() { return highEdProp; }
+  double getGas() { return gas; }
+  double getSewage() { return sewage; }
+  std::vector<double> getAgeProp() { return ageProp; }
+  double getMaleProp() { return maleProp; }
+  double getHighEdProp() { return highEdProp; }
   double getBirthRate() { return birthRate; }
   double getDeathRateAdults() { return deathRateAdults; }
   double getDeathRateAquatic() { return deathRateAquatic; }
   std::list<Human> getHumans() { return humans; }
   std::list<Mosquito> getMosquitoes() { return mosquitoes; }
-  std::vector<float> getMovementPatterns() { return movementPatterns; }
+  std::vector<double> getMovementPatterns() { return movementPatterns; }
 
-  Territory(int, float, float, int, const std::vector<float>&, float, float,
-            float, float, const std::vector<float>&);
+  Territory(int, double, double, int, const std::vector<double>&, double, double,
+            double, double, const std::vector<double>&);
   Territory();
 
-  void initializeHumans(float, float, float);
-  void initializeMosquitoes(int, float, bool);
+  void initializeHumans(double, double, double);
+  void initializeMosquitoes(int, double, bool);
   void addHuman(Human);
   void removeHuman(Human);
   void resetHumans();
   void addMosquito(Mosquito);
   void removeMosquito(Mosquito);
-  void updateDeathRates(float, int);
-  void updateBirthRate(float, int);
+  void updateDeathRates(double, int);
+  void updateBirthRate(double, int);
   void moveHumans();
   void updateMosquitoes();
   void updateHumans(int);
   void interaction(int);
   bool checkProximity(Human, Mosquito);
   void moveMosquitoes();
-  void birthMosquitoes(float, float);
-  void deathMosquitoes(float, float);
+  void birthMosquitoes(double, double);
+  void deathMosquitoes(double, double);
   std::vector<int> contagions();
-  int weightedRandom(std::vector<float>);
+  int weightedRandom(std::vector<double>);
   std::list<Human> humans;
   std::list<Mosquito> mosquitoes;
 
  private:
   int id;
-  float area;
+  double area;
   int width;
   int length;
-  float density;
+  double density;
   int population;
-  std::vector<float> ageProp;
-  float maleProp;
-  float highEdProp;
-  float gas;
-  float sewage;
+  std::vector<double> ageProp;
+  double maleProp;
+  double highEdProp;
+  double gas;
+  double sewage;
   double birthRate;
   double deathRateAdults;
   double deathRateAquatic;
-  std::vector<float> movementPatterns;
+  std::vector<double> movementPatterns;
 };
 
 #endif  // SRC_TERRITORY_H_
